task-1 v03

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -1,10 +1,11 @@
 package com.epam.utils;
 
 public class StringUtils {
+
     public static boolean isPositiveNumber(String str) {
          //here magic will happen
         try {
-            int number = Integer.parseInt(str);
+            double number = Double.parseDouble(str);
             return number > 0;
         } catch (NumberFormatException e) {
             return false;
